Gnomonic deterministic

Estimate natural mortality based on gnomonic interval approach.

gnomonic(
  nInterval,
  eggDuration,
  addInfo = NULL,
  longevity,
  fecundity,
  a_init = 2
)

Arguments

nInterval
a numeric value that represents the number of gnomonic intervals.

eggDuration
a numeric value with the egg stage (first gnomonic interval) duration in days.

addInfo
a numeric vector with additional information related to the observed duration of the other gnomonic intervals different than the first interval (egg stage duration). Write addInfo = NULL if you do not provide additional information.

longevity
a numeric value indicating the lifespan of the species in days.

fecundity
a numeric value indicating the mean lifetime fecundity (MLF) as the number of eggs produced for a female.

a_init
a numeric value indicating the initial parameter related to the proportionality optimized by iterative solution via univariate (1-dim.) minimization. a_init = 2 as default value.

Value

A list of class 'gnomos'.

a the proportionality constant.

G the constant proportion of the overall natural death rate.

results a dataframe with the duration ("interval_duration_day") and natural mortality ("M_day" and "M_year") for each gnomonic interval.

Details

Estimate natural mortality (M) based on gnomonic interval approach.

The argument nInterval is NULL by default. If you have -at least- one observed value for the duration of the other gnomonic intervals you should provide this as a vector which length must be nInterval - 1, for example addInfo = c(3, NA, NA, NA, NA, NA)) for a nInterval = 7.

References

Caddy JF (1996). Modelling natural mortality with age in short-lived invertebrate populations: definition of a strategy of gnomonic time division. Aquat Living Resour 9:197–207. doi: 10.1051/alr:1996023.

Examples

#The values are based on Caddy (1996).
model <- gnomonic(nInterval = 7, eggDuration = 2, addInfo = NULL,
longevity = 365, fecundity = 200000, a_init = 2)

#> -------------------------------------------------------- 
#> 
#> No additional information. You are only considering the egg stage duration = 2 
#> 
#> -------------------------------------------------------- 
#> 

model

#> Proportionality constant (alpha) = 1.381646 
#> 
#> -------------------------------------------------------- 
#> 
#> Constant proportion of the overall natural death rate (G) = 1.644704 
#> 
#> -------------------------------------------------------- 
#> 
#> Main results of gnomonic method: 
#> 
#>   Gnomonic_interval interval_duration_day total_duration M_day  M_year No_Surv
#> 1                 1                 2.000              2 0.822 300.158   38614
#> 2                 2                 2.763              5 0.595 217.247    7455
#> 3                 3                 6.581             11 0.250  91.217    1439
#> 4                 4                15.674             27 0.105  38.300     278
#> 5                 5                37.330             64 0.044  16.081      54
#> 6                 6                88.907            153 0.018   6.752      10
#> 7                 7               211.745            365 0.008   2.835       2

model$a

#> [1] 1.381646

model$G

#> [1] 1.644704

model$results

#>   Gnomonic_interval interval_duration_day total_duration M_day  M_year No_Surv
#> 1                 1                 2.000              2 0.822 300.158   38614
#> 2                 2                 2.763              5 0.595 217.247    7455
#> 3                 3                 6.581             11 0.250  91.217    1439
#> 4                 4                15.674             27 0.105  38.300     278
#> 5                 5                37.330             64 0.044  16.081      54
#> 6                 6                88.907            153 0.018   6.752      10
#> 7                 7               211.745            365 0.008   2.835       2

#Additional information for the duration of the second gnomonic intervals.
model <- gnomonic(nInterval = 7, eggDuration = 2, addInfo = c(3, NA, NA, NA, NA, NA),
longevity = 365, fecundity = 200000, a_init = 2)

model

#> Proportionality constant (alpha) = 1.381388 
#> 
#> -------------------------------------------------------- 
#> 
#> Constant proportion of the overall natural death rate (G) = 1.644704 
#> 
#> -------------------------------------------------------- 
#> 
#> Main results of gnomonic method: 
#> 
#>   Gnomonic_interval interval_duration_day total_duration M_day  M_year No_Surv
#> 1                 1                 2.000              2 0.822 300.158   38614
#> 2                 2                 3.000              5 0.548 200.106    7455
#> 3                 3                 6.579             12 0.250  91.244    1439
#> 4                 4                15.668             27 0.105  38.316     278
#> 5                 5                37.311             65 0.044  16.090      54
#> 6                 6                88.852            153 0.019   6.756      10
#> 7                 7               211.590            365 0.008   2.837       2

model$a

#> [1] 1.381388

model$G

#> [1] 1.644704

model$results

#>   Gnomonic_interval interval_duration_day total_duration M_day  M_year No_Surv
#> 1                 1                 2.000              2 0.822 300.158   38614
#> 2                 2                 3.000              5 0.548 200.106    7455
#> 3                 3                 6.579             12 0.250  91.244    1439
#> 4                 4                15.668             27 0.105  38.316     278
#> 5                 5                37.311             65 0.044  16.090      54
#> 6                 6                88.852            153 0.019   6.756      10
#> 7                 7               211.590            365 0.008   2.837       2