Print method for gnomosBoot class

Print method for gnomosBoot class

# S3 method for gnomosBoot
print(x, ...)

Arguments

x
an object class 'gnomosBoot'.

...
Additional arguments to the print method.

Value

The values of the proportionality constant (a), constant proportion of the overall natural death rate (G) and a data.frame with the duration and natural mortality for each gnomonic interval.

Examples

#The values are based on Caddy (1996).
modelBoot <- gnomonicStochastic(nInterval = 7, eggDuration = 2, addInfo = NULL, longevity = 365,
distr = "uniform", min_fecundity = 100000, max_fecundity = 300000, niter = 50, a_init = 2)

#> -------------------------------------------------------- 
#> 
#> No additional information. You are only considering the egg stage duration = 2 
#> 
#> -------------------------------------------------------- 
#> 
#> [1] "You are using a 'uniform distribution' for fecundity."

print(modelBoot)

#> Proportionality constant (alpha) = 1.381646 
#> 
#> -------------------------------------------------------- 
#> 
#> Mean value of constant proportion of the overall natural death rate (G) = 1.647106 
#> 
#> -------------------------------------------------------- 
#> 
#> M values in 1/day for each gnomonic interval for each iteration: 
#> 
#>              Gnomonic_1 Gnomonic_2 Gnomonic_3 Gnomonic_4 Gnomonic_5 Gnomonic_6
#> Iteration_1   0.8464460  0.6126361  0.2572323 0.10800610 0.04534936 0.01904119
#> Iteration_2   0.8261556  0.5979505  0.2510661 0.10541707 0.04426228 0.01858475
#> Iteration_3   0.8511038  0.6160073  0.2586478 0.10860045 0.04559891 0.01914597
#> Iteration_4   0.8448716  0.6114966  0.2567538 0.10780522 0.04526501 0.01900577
#> Iteration_5   0.8395589  0.6076514  0.2551393 0.10712732 0.04498038 0.01888626
#> Iteration_6   0.8361608  0.6051919  0.2541066 0.10669372 0.04479832 0.01880982
#> Iteration_7   0.7844734  0.5677819  0.2383990 0.10009844 0.04202911 0.01764709
#> Iteration_8   0.8444580  0.6111973  0.2566281 0.10775244 0.04524285 0.01899647
#> Iteration_9   0.8408456  0.6085827  0.2555303 0.10729150 0.04504931 0.01891520
#> Iteration_10  0.8233263  0.5959027  0.2502063 0.10505604 0.04411070 0.01852110
#> Iteration_11  0.8211736  0.5943446  0.2495521 0.10478136 0.04399536 0.01847267
#> Iteration_12  0.8344776  0.6039737  0.2535951 0.10647894 0.04470814 0.01877195
#> Iteration_13  0.8388407  0.6071316  0.2549211 0.10703568 0.04494190 0.01887010
#> Iteration_14  0.7982500  0.5777531  0.2425857 0.10185632 0.04276720 0.01795700
#> Iteration_15  0.7965574  0.5765280  0.2420713 0.10164034 0.04267652 0.01791892
#> Iteration_16  0.8338221  0.6034993  0.2533959 0.10639531 0.04467302 0.01875721
#> Iteration_17  0.8415103  0.6090638  0.2557323 0.10737631 0.04508493 0.01893016
#> Iteration_18  0.7988516  0.5781885  0.2427685 0.10193308 0.04279943 0.01797053
#> Iteration_19  0.8387394  0.6070583  0.2548903 0.10702275 0.04493647 0.01886783
#> Iteration_20  0.8454179  0.6118920  0.2569198 0.10787492 0.04529428 0.01901806
#> Iteration_21  0.8309578  0.6014262  0.2525255 0.10602982 0.04451956 0.01869277
#> Iteration_22  0.8056095  0.5830797  0.2448222 0.10279539 0.04316150 0.01812255
#> Iteration_23  0.7862290  0.5690526  0.2389325 0.10032245 0.04212316 0.01768658
#> Iteration_24  0.8452073  0.6117395  0.2568558 0.10784804 0.04528299 0.01901332
#> Iteration_25  0.8502587  0.6153956  0.2583909 0.10849260 0.04555363 0.01912696
#> Iteration_26  0.7732324  0.5596459  0.2349829 0.09866408 0.04142685 0.01739421
#> Iteration_27  0.8124759  0.5880494  0.2469089 0.10367154 0.04352937 0.01827701
#> Iteration_28  0.8235832  0.5960886  0.2502843 0.10508883 0.04412446 0.01852688
#> Iteration_29  0.7979043  0.5775028  0.2424806 0.10181221 0.04274868 0.01794922
#> Iteration_30  0.8449562  0.6115578  0.2567795 0.10781601 0.04526954 0.01900767
#> Iteration_31  0.8452851  0.6117959  0.2568795 0.10785798 0.04528717 0.01901507
#> Iteration_32  0.8439731  0.6108463  0.2564808 0.10769057 0.04521688 0.01898556
#> Iteration_33  0.8472754  0.6132364  0.2574843 0.10811194 0.04539380 0.01905985
#> Iteration_34  0.8020563  0.5805080  0.2437424 0.10234200 0.04297113 0.01804262
#> Iteration_35  0.7996786  0.5787870  0.2430198 0.10203860 0.04284374 0.01798913
#> Iteration_36  0.7955730  0.5758155  0.2417721 0.10151473 0.04262378 0.01789678
#> Iteration_37  0.8196404  0.5932349  0.2490861 0.10458573 0.04391322 0.01843818
#> Iteration_38  0.7838740  0.5673481  0.2382169 0.10002196 0.04199699 0.01763360
#> Iteration_39  0.8394271  0.6075560  0.2550993 0.10711051 0.04497332 0.01888330
#> Iteration_40  0.8286510  0.5997566  0.2518244 0.10573548 0.04439597 0.01864088
#> Iteration_41  0.8178747  0.5919569  0.2485495 0.10436042 0.04381862 0.01839846
#> Iteration_42  0.8070173  0.5840986  0.2452500 0.10297502 0.04323692 0.01815422
#> Iteration_43  0.7971855  0.5769826  0.2422622 0.10172049 0.04271017 0.01793305
#> Iteration_44  0.8470251  0.6130553  0.2574083 0.10808000 0.04538039 0.01905422
#> Iteration_45  0.8370767  0.6058549  0.2543850 0.10681059 0.04484739 0.01883042
#> Iteration_46  0.8338012  0.6034841  0.2533895 0.10639263 0.04467190 0.01875674
#> Iteration_47  0.8442636  0.6110566  0.2565691 0.10772764 0.04523244 0.01899209
#> Iteration_48  0.8009151  0.5796820  0.2433956 0.10219638 0.04290999 0.01801695
#> Iteration_49  0.8443624  0.6111281  0.2565991 0.10774025 0.04523773 0.01899432
#> Iteration_50  0.7872382  0.5697830  0.2392392 0.10045122 0.04217723 0.01770928
#>               Gnomonic_7
#> Iteration_1  0.007994971
#> Iteration_2  0.007803322
#> Iteration_3  0.008038966
#> Iteration_4  0.007980100
#> Iteration_5  0.007929921
#> Iteration_6  0.007897823
#> Iteration_7  0.007409619
#> Iteration_8  0.007976194
#> Iteration_9  0.007942073
#> Iteration_10 0.007776598
#> Iteration_11 0.007756265
#> Iteration_12 0.007881925
#> Iteration_13 0.007923137
#> Iteration_14 0.007539743
#> Iteration_15 0.007523756
#> Iteration_16 0.007875734
#> Iteration_17 0.007948352
#> Iteration_18 0.007545426
#> Iteration_19 0.007922180
#> Iteration_20 0.007985260
#> Iteration_21 0.007848680
#> Iteration_22 0.007609257
#> Iteration_23 0.007426201
#> Iteration_24 0.007983271
#> Iteration_25 0.008030983
#> Iteration_26 0.007303443
#> Iteration_27 0.007674112
#> Iteration_28 0.007779024
#> Iteration_29 0.007536478
#> Iteration_30 0.007980899
#> Iteration_31 0.007984006
#> Iteration_32 0.007971614
#> Iteration_33 0.008002805
#> Iteration_34 0.007575695
#> Iteration_35 0.007553237
#> Iteration_36 0.007514458
#> Iteration_37 0.007741783
#> Iteration_38 0.007403958
#> Iteration_39 0.007928676
#> Iteration_40 0.007826891
#> Iteration_41 0.007725105
#> Iteration_42 0.007622553
#> Iteration_43 0.007529689
#> Iteration_44 0.008000441
#> Iteration_45 0.007906475
#> Iteration_46 0.007875536
#> Iteration_47 0.007974358
#> Iteration_48 0.007564916
#> Iteration_49 0.007975291
#> Iteration_50 0.007435733
#> -------------------------------------------------------- 
#> 
#> Main results of gnomonic method: 
#>   Gnomonic_interval duration total_duration     M_lower       M_day     M_upper
#> 1                 1    2.000              2 0.784008903 0.823552973 0.849587423
#> 2                 2    2.763              5 0.567445730 0.596066724 0.614909797
#> 3                 3    6.581             11 0.238257836 0.250275155 0.258186942
#> 4                 4   15.674             27 0.100039164 0.105084969 0.108406952
#> 5                 5   37.330             64 0.042004219 0.044122840 0.045517667
#> 6                 6   88.907            153 0.017636637 0.018526199 0.019111856
#> 7                 7  211.745            365 0.007405231 0.007778739 0.008024643
#>           M_sd     M_year No_Surv
#> 1 0.0222636595 300.596835   40957
#> 2 0.0161138713 217.564354    7889
#> 3 0.0067658560  91.350431    1519
#> 4 0.0028408324  38.356014     293
#> 5 0.0011928023  16.104837      56
#> 6 0.0005008312   6.762063      11
#> 7 0.0002102879   2.839240       2